Updates to use OAuth2 0.8.0 (#27)
* Updates OAuth2 deps.

* Updates Facebook token request method.

According to the Facebook's OAuth specifics exchanging a code for a
token is performed via a GET request. Changing this to a GET request
specifically ensures that the response type is JSON, so we can parse the
token, token_type, and expires_at parameters properly.

https://developers.facebook.com/docs/facebook-login/manually-build-a-login-flow#confirm

* Adds client_secret to token request params.

* Sets token errors properly.

* Uses Facebook API v2.8

Without specifying the API version the API was returning `text/plain` as
the mime type. Locking it here ensures that it returns JSON. We may want
to add this as a config option for the strategy.

* Relaxes oauth2 version dependency